19.2.Manually Profile (Instrument) Part of your Code{Installing the Blackfire SDK; Manually Instrumenting Code; How Disabling Auto-Instrumentation Works}
When composer require blackfire/php-sdk finishes, move over and open src/Controller/MainController.php.
Ok: this is the controller for our homepage. Let's pretend that when we profile this page, we don't want to collect data about all of our code. Nope, we want to, sort of, "zoom in" and see only what's happening inside the controller.

Manually Instrumenting Code
We can do that by saying $probe = \BlackfireProbe::getMainInstance(). Remember: the PHP extension is called the "probe"... that's important if you want this to make sense. Then call $probe->enable().
At the bottom, I'll set the rendered template to a $response variable, add $probe->disable() and finish with return $response.
Okay, so... what the heck does this do? The first thing I want you to notice is that if I refresh the homepage a bunch of times... and then go to https://blackfire.io, I do not have any new profiles. Adding this code does not "trigger" a new profile to be created: it does not tell the PHP extension - the "probe" - that it should to do its work.
Instead, if a profile is currently being created, this tells the probe when to start collecting data. Hmm, this isn't going to quite make sense until we see it in action. Trigger a new profile on the homepage. I'll call this one: [Recording] Only instrument some code.
Click to view the call graph: https://bit.ly/sf-bf-partial-profile.
Fascinating. This contains less information than normal. It has a few things on top - main() and handleRaw()... but basically it jumps straight to the homepage() method.

How Disabling Auto-Instrumentation Works
What's happening here is that the only code that the probe "instrumented", the only code that it collected information on, is the code between the enable() and disable() calls.
This... completely confused me the first time I saw it. What really happens is this: as soon as we use the browser extension to tell the probe to do its job, the PHP extension starts instrumenting - so, collection data - immediately. Initially, it is collecting data about every line of PHP code.
But as soon as it sees $probe->enable(), it basically forgets about all the data collected so far. The $probe->enable() call says:

  Hey! Start instrumenting here. If you've already collected some data before thanks to auto-instrumentation, get rid of it.

This effectively disables auto-instrumentation: we're now controlling which code is instrumented instead of it happening automatically. Once the code hits $probe->disable() instrumentation stops.
You can actually use $probe->enable() and $probe->disable() multiple times in your code if you want to profile different pieces: $probe->enable() only forgets data it's already collected the first time you call it.