21.1.Creating an Automatic Probe Early in your Code{Generating the Event Subscriber}
Once we determine that we want to create a probe dynamically in our code, we really want to do that as early as possible so that Blackfire can "instrument" as much of our code as possible.

Generating the Event Subscriber
In Symfony, we can do that with an event subscriber... which we will generate to be super lazy. First, in .env, make sure that you're back in the dev environment.

Then, find your terminal and run:

php bin/console make:subscriber
Call it BlackfireAutoProfileSubscriber... and we want to listen to RequestEvent: Go check out the code src/EventSubscriber/BlackfireAutoProfileSubscriber.php