21.2.Creating an Automatic Probe Early in your Code{Generating the Event Subscriber; Creating the Prove in the Subscriber}
So, when this RequestEvent happens - which Symfony dispatches super early when handling a request, we want to create & enable the probe. Copy all of the $shouldProfile code, remove it from the controller and paste it here.

Creating the Prove in the Subscriber
Now add $request = $event->getRequest(). To make this only profile the GitHub organization AJAX call - whose URL is /api/github-organization - set $shouldProfile equal to $request->getPathInfo() === '/api/github-organization'.
In a real app, I would add more code to make sure $shouldProfile is only true on the very specific requests we want to profile.
Now I'll re-type the t on Client and select the correct Client class so that PhpStorm adds that use statement to the top of the class for me.
Thanks PhpStorm!
But before we try this, I want to code for one edge case: if not $event->isMasterRequest(), then return.
It might not be important in your app, but Symfony has a "sub-request" system... and the short explanation is that we don't want to profile those: they are not real requests... and would make a big mess of things.
Ok, let's try this! I'll close a tab... then refresh the homepage... which causes the AJAX request to be made. You can see it's slow. Now reload the list of profiles on Blackfire... there it is! Open it up.
And... oh wow, oh weird! 281 microseconds. Give this a name: [Recording] Auto from subscriber: http://bit.ly/sf-bf-broken-auto-profile
This profile is... broken. That's 281 microseconds - so .281 milliseconds. And the entire profile is just the Probe::enable() call itself!