21.3.Creating an Automatic Probe Early in your Code{Prevent probe Auto-Close Too Early}
Probe Auto-Close Too Early
The solution is to call $probe->close() manually... which - more importantly - will require us to store the Probe object in a way where PHP won't garbage collect it too early.
So here's the goal: call $probe->close() as late as possible during the request lifecycle. We can do this by listening to a different event: when TerminateEvent::class is dispatched - that's very late in Symfony - call the onTerminateEvent() method.
I'll hit an Alt+Enter shortcut to create that method... then add the argument TerminateEvent $event.
To be able to call $probe->close(), we need to store the probe object on a property. Add private $probe with some documentation that says that this will either be a Probe instance from Blackfire or null. Update the code below to be $this->probe = $blackfire->createProbe().
Finally, inside onTerminateEvent, if $this->probe - I should not have put that exclamation point, that's a mistake - then $this->probe->close()
If you assume that I did not include the exclamation point... then this makes sense! If we created the probe, then we will close it. Problem solved. And... really... the fact that we set the probe onto a property is the real magic: that will prevent PHP from garbage-collecting that object... which will prevent it from closing itself until we're ready.