21.4.Creating an Automatic Probe Early in your Code{Increasing the Event Priority}
Increasing the Event Priority
While we're here, let's make this a little bit cooler. Change onRequestEvent to be an array... and add 1000 as the second item.
This syntax is... weird. But the result is good: it says that we want to listen to this event with a priority of 1000. That will make our code run even earlier so that even more code will get profiled.