22.1.Performance Tests{We have a Performance "Bug"; Running the Functional Test}
Let's profile the Github API endpoint again. I'll cheat and go directly to /api/github-organization... and click to profile this. I'll call it: [Recording] GitHub Ajax HTTP requests because we're going to look closer at the HTTP requests that our app makes to the GitHub API.
Click to view the call graph: https://bit.ly/sf-bf-http-requests
Oh wow - this request was super slow - 1.83 seconds - a lot slower than we've seen before. We can see that curl_multi_select() is the problem: this is our code making requests to the GitHub API, which is apparently running a bit slow at the moment.

We have a Performance "Bug"
Lucky for us, that's exactly what I wanted to talk about! At the top, Blackfire tells me that this page made two HTTP requests. And HTTP requests are always expensive for performance.
If you studied the data from the two API endpoints that we're using, you would discover that it's possible - by writing some clever code - to get all the info our app needs with just one HTTP request.
What I'm saying is: our page is making one more HTTP request than it truly needs to. If you think about it... that's kind of a performance "bug": we're making 2 HTTP requests and we only need 1.
In an ideal world, when we find a bug, the process for fixing it looks like this. First, write a test for the expected behavior. Second, run that test and watch it fail. And third, fix the bug and make sure the test passes.
Whelp, when it comes to a performance bug... we can do the exact same thing! We can write a functional test that asserts that this endpoint only makes one HTTP request. It's... pretty awesome.

Running the Functional Test
Find your editor and open tests/Controller/MainControllerTest.php. I already set up a functional test that makes a request to /api/github-organization and checks some basic data on the response:

 19 lines  tests/Controller/MainControllerTest.php
namespace App\Tests\Controller;
use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
class MainControllerTest extends WebTestCase
{
    public function testGetGitHubOrganization()
    {
        $client = static::createClient();
        $client->request('GET', '/api/github-organization');
        $this->assertResponseIsSuccessful();
        $data = json_decode($client->getResponse()->getContent(), true);
        $this->assertArrayHasKey('organization', $data);
    }
}
Let's makes sure this passes. Run PHPUnit and point it directly at this class:

php bin/phpunit tests/Controller/MainControlerTest.php
The first time you run this script, it will probably download PHPUnit in the background. When it finishes... go tests go! All green.