22.2.Performance Tests{Adding a Performance Assertion}
Adding a Performance Assertion
Here's the idea: in addition to asserting that this response contains JSON with an organization key, I also want to assert that it only made one HTTP request. To do that, first add a trait from the SDK: use TestCaseTrait. Next, in the method, add $blackfireConfig = new Configuration() - the one from Blackfire\Profile: the same Configuration class we used earlier when we gave our custom-created profile a title. This time call assert() and pass it a very special string: metrics.http.requests.count == 1.
I'll show you where that came from soon. Finally, below this, call $this->assertBlackfire() and pass this $blackfireConfig and a callback function.
So... this confused me at first. When we call $this->assertBlackfire() it will execute this callback. Inside, we will do whatever work we want - like making the request. Finally, when the callback finishes, Blackfire will execute this assertion against the code that we ran.
To get this to work, we need to use ($client).
If this doesn't make sense yet... don't worry: we'll dive a bit deeper soon. But right now... try it! Run the test again:

php bin/phpunit tests/Controller/MainControlerTest.php
And... it fails! Woo! Failed that metrics.http.requests.count == 1!

Performance Tests Create Real Profiles
Behind the scenes, the Blackfire SDK created a real Blackfire profile for the request! You can even copy the profile URL and go check it out! This takes us to an "assertions" tab. We're making 2 requests instead of the expected one. We'll talk a lot more about assertions soon.
Ok, but how did this really work? It's beautifully simple. When you run the test, it does make a real Blackfire profile in the background. However, if you go to your Blackfire homepage, you won't see it.
Why? Hold Cmd or Ctrl and click the assertBlackfire() method. I love it: this method uses the SDK - just like we did! - to create a real profile. When it does that, it also adds a skip_timeline option, which simply tells Blackfire to hide this from our profile page... so it doesn't get cluttered up with all these test profiles. You can totally override that if you wanted... via the Configuration object.
In reality, the Blackfire PHPUnit integration is doing the exact same thing that we just finished doing in our code: manually creating a new profile. This is really nothing new... and I love that!
Except... for this metrics thing. Where did that string come from? And what else can we do here? Let's dive into metrics next.