23.1.All about Metrics{What Metrics are Available}
Where did this metrics string come from - this metrics.http.requests.count?

 29 lines  tests/Controller/MainControllerTest.php
class MainControllerTest extends WebTestCase
{
    public function testGetGitHubOrganization()
    {
        $blackfireConfig = (new Configuration())
            ->assert('metrics.http.requests.count == 1');
    }
}
There are two things I want to say about this. First, Blackfire stores tons of raw data about your profile in little "categories" called metrics. More on that soon. And second, inside the assert() call, you're using a special "expression" language that's similar to JavaScript. It's technically Symfony's ExpressionLanguage if you want to read more. Behind-the-scenes, metrics is probably some object... and we're referencing an http property, then a requests... property then a count property & then we're comparing that to 1.

What Metrics are Available
Ok, cool. So... how the heck did I know to use this exact string to get the HTTP call count? This goes back to the Blackfire timeline. On the profile, click the timeline link.
When we talked about the timeline earlier, we talked about how, on the left side, there are these "timeline" metrics. At that point, these were just a nice way to add color to different sections of the timeline.
But now we understand that there is a lot more power behind this info: this shows us all the pieces of data we can use in our tests... and in other places that we'll talk about soon.
For example, there's a metric called symfony.events.count which equals seven. You could use that in a metric if, for some reason, you wanted to assert that a certain number of events were dispatched. If I needed to do an assertion about the number of HTTP requests, I would probably search the metrics for http. Apparently there are two... and if you looked closer, you'd find that http.requests is perfect. Most of these metrics have data about multiple dimensions: we can say http.requests.count to get the actual number or http.requests.memory to get how much memory they used.
In the test system, we start with metrics. then use anything we find here.