23.2.All about Metrics{Fixing the Performance Bug}
Fixing the Performance Bug
We now have a performance bug in our application that we've proven with a test. And at this point, the actual way we fix that bug is not as important: all we care about is that we can change some code and get this test to pass.
The logic for the API calls lives in src/GitHub/GitHubApiHelper.php: it has two public function and each makes one API request.
How can we make this page only make 1 HTTP request? Well, if you looked closely.. Ah! Too close! Ahh. You'd find that you can get all the information you need by only making this second HTTP request. The details aren't important - so let's just jump in.
Add a new property called $githubOrganizations set to an empty array.
As we loop over the repositories for a specific organization, we will store that organization's info. Add a new variable called $publicRepoCount set to 0 the number of public repositories an organization has is one of the pieces of data we need.

Then, inside the foreach: if $repoData['private'] === false - that's one of the keys on $repoData - say ++$publicRepoCount.
So, as we're looping over the repositories, we're counting how many are public.
Finally, at the bottom, if not isset($this->githubOrganizations[$organization]), then $this->githubOrganizations[$organization] = new GitHubOrganization().
This needs two arguments. The first is the organization name. We can probably use the $organization argument... or you can use $data[0] - to get the first repository - then ['owner']['login']. For the second argument, pass $publicRepoCount.
Now, each time we call this method, we capture the organization's information and store it on this property. So if we call this method first and then the other method... we could cheat and return the GitHubOrganization object that's stored on the property. It's property caching!
Check it out: if isset($this->githubOrganizations[$organization]) then return that immediately without doing any work.
So... are we calling these two methods in the "correct" order to get this to work? Check out the controller. Nope! Swap these two lines so the first call will set up the caching for the second.

Phew! Let's see if that helps. It was a complicated fix... but thanks to our test, we will know for sure if it worked. Go!

php bin/phpunit tests/Controller/MainControlerTest.php
They pass! This proves that we reduced the HTTP calls from two to one.