23.3.All about Metrics{Typos in Metrics}
Typos in Metrics
What I love about the metrics system is that there are many to choose from. What I don't love is that you need to manually look up everything that's available. Fortunately, if you make a typo - the error is great. Change count to vount.
And re-run the test:

php bin/phpunit tests/Controller/MainControlerTest.php
  An error occurred when profiling the test

And when we follow the profile link... check out that error!

 The following assertions are not valid... Property "vount" does not exist, available ones are:

... and it lists all the properties.