23.4.All about Metrics{Organizing Blackfire Assertions into Separate Test Cases; Don't do Time-Based Assertions}
Organizing Blackfire Assertions into Separate Test Cases
The one downside to adding Blackfire assertions in your tests is that they do slow things down because instrumentation happens and we need to wait for Blackfire to create the profile.
Because of that, as a best practice, we usually like to isolate our performance tests from our normal tests. Check it out: copy the test method name, paste it below, and call it testGetGitHubOrganizationBlackfireHttpRequests().
And... copy the contents of the original method and paste here. Now... we only need to create the $client, create $blackfireConfig and, inside assertBlackfire(), just make the request.
Back in the original method, we can simplify... in fact we can go all the way back to the way it was before: create the client, make the request, assert something.
Why is this useful? Because now we can skip the Blackfire tests if we're just trying to get something to work. How? Above the performance test, add @group blackfire.

Thanks to that, we can add --exclude-group=blackfire to avoid the Blackfire tests:

 php bin/phpunit tests/Controller/MainControlerTest.php --exclude-group=blackfire
Yep! Just one test, two assertions. Another nice detail is to add @requires extension blackfire.
Now, if someone is missing the Blackfire extension, instead of the tests exploding, they'll be marked as skipped.

Don't do Time-Based Assertions
The last thing I want to mention about assertions is this: please, please please avoid time-based assertions. They're the easiest to create - I know. It's super tempting to want to create an assertion that the request should take less than 500 milliseconds. If you do this, you will hate your tests.
Why? Because there's way too much variability in time: the request might run fast enough on one machine, but not fast enough on another. Or your server might just have a bad day... and suddenly your tests are failing. Relying on time makes your tests fragile.
Next, we're going to talk more about metrics and assertions. We know that we can add assertions to profiles that are created inside our tests.

But we an also add global assertions: tests that run any time you create a profile for any page! If you want to make sure that a specific page - or any page - doesn't make more than, I don't know, 10 database queries, you can add an assertion for that and see a big failure if you break the rules.