26.1.Expectations/Tests with Blackfire Player{Adding an Expectation/Test to a Page; What's Possible in the expect Expression?}
We just used blackfire-player to execute our first "scenario". It's pretty simple: it goes to the homepage then clicks the "Log In" link.
It works... but... we're not doing anything after we visit these pages. The true power of blackfire-player is that you can add tests to your scenario - or even scrape pages and save that data somewhere.

Adding an Expectation/Test to a Page
To add a "test" - or "assertion", or "expectation"... I love when things have 5 names... - say expect followed by - you guessed it! - an expression! status_code() == 200. Copy that and add it to the login page as well.
Ok, try blackfire-player again!

blackfire-player run scenario.bkf --ssl-no-verify -v
Woo! It still passes and now it's starting to be useful!

What's Possible in the expect Expression?
Let's break this down. First, just like we saw with the metrics stuff:

 40 lines  tests/Controller/MainControllerTest.php
class MainControllerTest extends WebTestCase
{
    public function testGetGitHubOrganizationBlackfireHttpRequests()
    {
        $blackfireConfig = (new Configuration())
            ->assert('metrics.http.requests.count == 1');
    }
}
This is an expression - it's Symfony's ExpressionLanguage once again - basically JavaScript. And second... this expression has a ton of built-in functions.
Search the blackfire-player docs for "status_code"... and keep searching until you find a big function list. Here it is. Yep, we can use current_url(), header() to get a header value and many others. The css() function is especially useful: it allows us to dig into the HTML on the page. We'll use that in a minute. The docs also have good examples of how to do more complex things. But we're not going to become Blackfire player experts right now... I just want you to get comfortable with writing scenarios.