26.3.Expectations/Tests with Blackfire Player{Typos in Expressions; Performance Assertions in the Scenarios?}
Typos in Expressions
Another thing that blackfire-player does well is its errors when I... do something silly. Make a typo: change count() to ount().
And rerun the scenario:

blackfire-player run scenario.bkf --ssl-no-verify -v
  Unable to call method ount of object Crawler.

That's a huge hint to tell you what object you're working with so you can figure out what methods it does have. Change that back to count().

Performance Assertions in the Scenarios?
So... blackfire-player has nothing to do with the Blackfire profiler. It's just a useful tool for visiting pages, clicking on links and adding expectations. But... if it truly had nothing to do with the profiler, I probably wouldn't have talked about it. In reality, the concept of "scenarios" is about to become very important - it's a fundamental part of a topic we'll talk about soon: Blackfire "builds".
And actually, there is one little integration between blackfire-player and the profiler: you can add performance assertions to your scenario. To do that, instead of expect, say assert and then use any performance expression you want: the same strings that you can use inside a test. For example: metrics.sql.queries.count < 30.
When we execute this:

blackfire-player run scenario.bkf --ssl-no-verify -v
It does still pass.