26.4.Expectations/Tests with Blackfire Player{Performance Assertions in the Scenarios?}
But if you played with this value - like set it to < 1 and re-ran the scenario:

blackfire-player run scenario.bkf --ssl-no-verify -v
Hmm, it still passes... even though this page is definitely making more than one query. The reason is that the assert functionality won't work inside a scenario until we introduce Blackfire "environments" - which we will soon. They are one of my absolute favorite parts of Blackfire.